+++ reactor-poller-core/src/main/java/com/jcarrey/reactor/poller/core/concurrency/MaxConcurrencyControlFn.java
@@ -3,7 +3,7 @@ package com.jcarrey.reactor.poller.core.concurrency;
 class MaxConcurrencyControlFn implements ConcurrencyControlFunction {
 
     @Override
-    public int calculateDelta(int currentConcurrency, ConcurrencyControlOperation operation) {
-        return Integer.MAX_VALUE;
+    public double calculateDelta(double currentConcurrency, ConcurrencyControlOperation operation) {
+        return Double.MAX_VALUE;
     }
 }
